requirements.txt
###############################################
# Core scientific stack
###############################################
numpy>=1.26
pandas>=2.1
pyarrow>=14.0
fastparquet>=2023.8.0

###############################################
# Fetching market data
###############################################
yfinance>=0.2.40
requests>=2.31.0

###############################################
# YAML config loading
###############################################
pyyaml>=6.0

###############################################
# Calendar + holiday features
###############################################
holidays>=0.34
pandas_market_calendars>=4.3

###############################################
# Machine learning utilities
###############################################
scikit-learn>=1.3
scipy>=1.11

###############################################
# PyTorch + Lightning + TFT
###############################################
torch>=2.2.0
pytorch-lightning>=2.1.3
pytorch-forecasting>=1.0.0
torchmetrics>=1.2.0

###############################################
# NLP (FinBERT sentiment)
###############################################
transformers>=4.39.0
tokenizers>=0.15.0

###############################################
# Filesystem + compression
###############################################
fsspec>=2023.10.0
zstandard>=0.22.0

###############################################
# Plotting + analysis
###############################################
matplotlib>=3.8
seaborn>=0.13

###############################################
# Utilities
###############################################
tqdm>=4.66
rich>=13.7
